fix: linting changes and refactor
