+++ internal/controllers/addon/phase_observe_operatorresource.go
@@ -116,3 +116,13 @@ func findConcernedCSVReference(csv client.ObjectKey, operator *operatorsv1.Opera
 	}
 	return nil
 }
+
+func addonUpgradeConcluded(addon *addonsv1alpha1.Addon, currentCSV client.ObjectKey, phase operatorsv1alpha1.ClusterServiceVersionPhase) bool {
+	// Upgrading has concluded if a new CSV(compared to the last known available) has come up and is
+	// in the succeeded phase.
+	if addonUpgradeStarted(addon) {
+		return currentCSV.String() != addon.Status.LastObservedAvailableCSV &&
+			phase == operatorsv1alpha1.CSVPhaseSucceeded
+	}
+	return false
+}
